Parallel jobs: get cost data from dynamic nodes
Previously, the LF's job cost scripts could not take into account the
agents that are dynamically allocated when an "agent" block is used
within a parallel stage. A new function has been added to the LF's
pipeline library that is intended to be run in a post block after any
parallel stage that uses a new agent.

This change also includes some minor whitespace cleanup in affected
files, as well as adding a license to edgeXGeneric and updating
copyright dates.

Issue: LF-Jira RELENG-3207
Signed-off-by: Eric Ball <eball@linuxfoundation.org>